feat: Add auto fill-in for email when navigating to /verify from /register
