+++ front/sssscs/src/verify/VerifyService.ts
@@ -1,0 +1,13 @@
+export enum VerificationMethod {
+    EMAIL, SMS
+};
+
+export interface VerificationCodeSendRequestDTO {
+    userEmail: string,
+    method: VerificationMethod,
+    dontActuallySend: boolean
+};
+
+export interface VerifyPageRouterState {
+    email: string
+}
